Move taskman web app features to a separate component.
