+++ src/TaskManagementWeb/App_Start/Startup.cs
@@ -1,13 +1,10 @@
-﻿using SenseNet.TaskManagement.Data;
-using SenseNet.TaskManagement.Hubs;
+﻿using SenseNet.TaskManagement.Hubs;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using SenseNet.Diagnostics;
 using SenseNet.Extensions.DependencyInjection;
-using EventId = SenseNet.Diagnostics.EventId;
 
 namespace SenseNet.TaskManagement.Web
 {
@@ -24,12 +21,6 @@ namespace SenseNet.TaskManagement.Web
         {
             services.AddControllersWithViews();
             services.AddSignalR();
-
-            SnLog.Instance = new SnFileSystemEventLogger();
-            SnTrace.SnTracers.Add(new SnFileSystemTracer());
-            SnTrace.EnableAll();
-
-            services.Configure<TaskManagementConfiguration>(Configuration.GetSection("TaskManagement"));
             
             //TODO: inject allowed origins dynamically (do not allow everything)
             services.AddCors(c =>
@@ -42,19 +33,16 @@ namespace SenseNet.TaskManagement.Web
                 });
             });
 
-            services.AddSingleton<TaskDataHandler>();
-            services.AddSenseNetClientTokenStore();
-            services.AddSingleton<ApplicationHandler>();
-            services.AddHostedService<DeadTaskHostedService>();
+            services.AddSenseNetTaskManagementWebServices(Configuration);
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ApplicationHandler appHandler)
         {
+            //var logger = app.ApplicationServices.GetService<ILogger<Program>>();
+
             // This will set the global SnLog and SnTrace instances to route log messages to the
             // official .Net Core ILogger API.
             app.ApplicationServices.AddSenseNetILogger();
-
-            SnLog.WriteInformation("Starting TaskManagement.Web", EventId.TaskManagement.Lifecycle);
 
             // make Web API use the standard ASP.NET error configuration
             //ConfigureErrorHandling();
@@ -99,8 +87,6 @@ namespace SenseNet.TaskManagement.Web
             
             //TODO: configure global error logging
             //httpConfiguration.Services.Add(typeof(IExceptionLogger), new WebExceptionLogger());
-            
-            SnLog.WriteInformation("SenseNet TaskManagement app started.", EventId.TaskManagement.Lifecycle);
 
             // load apps
             appHandler.Initialize();
