--- src/TaskManagementWeb/EventId.cs
@@ -1,18 +1,0 @@
-﻿//using System;
-//using System.Collections.Generic;
-//using System.Linq;
-//using System.Web;
-
-//namespace SenseNet.TaskManagement.Web
-//{
-//    internal static class EventId
-//    {
-//        internal static readonly int CommunicationError = 30;
-//        internal static readonly int DataError = 50;
-//        internal static readonly int CallbackError = 100;
-//        internal static readonly int AppLifecycle = 200;
-//        internal static readonly int TasksManagementError = 300;
-
-//        internal static readonly int InvalidUrl = 400;
-//    }
-//}
