+++ src/TaskManagementWeb/README.md
@@ -3,9 +3,3 @@
 Task Management Web is an ASP.NET web application. This is the central hub for managing tasks and notifying clients about the execution.
 
 It does not need much resources (as the tasks are performed by the agent machines), only an average web server and a database for storing task records and keeping track of their execution.
-
-For details and extensibility points please visit the following articles.
-
-- [Task Management overview](http://wiki.sensenet.com/Task_Management)
-- [Task Management for Developers](http://wiki.sensenet.com/Task_Management_-_for_Developers)
-- [Deployment](http://wiki.sensenet.com/Task_Management_deployment)
