--- src/TaskManagementWeb/TaskManagement/TaskExecutors/readme.txt
@@ -1,1 +1,0 @@
-﻿Place of task executor applications.
