--- src/TaskManagementWeb/credits.txt
@@ -1,20 +1,0 @@
-﻿-------------------------------------------------------------------
-sensenet - Enterprise Content Management for ASP.NET
--------------------------------------------------------------------
-CREDITS
--------------------------------------------------------------------
-
-These Sense/Net employees have helped sensenet Task Management 
-become what it is today. The list is in alphabetical order.
-
--------------------------------------------------------------------
-
-Anikó Litványi
-Anna Balog
-Ildikó Tamási
-Lajos Gallay
-Miklós Tóth
-Sándor Kiss
-Zoltán Gyebrovszki
-
-------------------------------END----------------------------------
