--- src/TaskManagementWeb/index.html
@@ -1,13 +1,0 @@
-﻿<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-    <title>sensenet Task Management</title>
-    <link rel="stylesheet" type="text/css" href="Content/Styles/main.css" />
-</head>
-<body>
-    <div class="tm-maintitle">
-        <img class="tm-mainimage" src="Content/Images/sn-smiley.png" />
-        <h1>Welcome to sensenet Task Management</h1>
-    </div>
-</body>
-</html>
